feat: add onCellChange callback and 'correct' classname on clues
As per the request in issue #20, added an **onCellChange** callback property.  I'm not sure in what
real-world cases this will help, but it's easy to add.  More importantly, Crossword now internally
tracks whether the guesses/answers for clues are correct or not, and passes that as a property
through `DirectionClues` and `Clue`, exposing a 'correct' class name on individual clues that are
correct.  This fixes the underlying need from issue #21.  The example app makes use of this by
styling "correct" clues with a checkmark and line-through, showing the player's progress through the
crossword.

re #20, fix #21