chore(dependencies): move snyk from a regular dependency to dev-only
Snyk provides install-time vulnerability protection, but that comes at a cost.  Some 185-or-so
additional dependencies get pulled in that react-crossword's consumers otherwise wouldn't need.
Plus, this forces Snyk's "patch vulnerabilities" opinion onto consumers, which is a lot for this
little component to demand.  (Consumers can still choose to use Snyk themselves, but that's *their*
choice, as it should be, rather than ours.)

re #33